examples/080-pipecat-voice-pipeline-python/requirements.txt
pipecat-ai[deepgram,daily,openai,silero]>=0.0.100
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
